Adds e2e test to send-message.js script and remove messages-content.js
