+++ tests/e2e/send-message.js
@@ -220,6 +220,20 @@ describe('Send message', () => {
     return name + ' - all contacts';
   };
 
+  const getElementText = (css, attempt) => {
+    attempt = attempt || 0;
+
+    return helper.getTextFromElement(element(by.css(css)))
+      .then((text) => {
+        return text;
+      }, (err) => {
+        if (attempt < 2) {
+          return getElementText(css, attempt+1);
+        }
+        throw err;
+      });
+  };
+
   describe('Send message modal', () => {
     it('can send messages to raw phone numbers', () => {
       common.goToMessages();
@@ -382,6 +396,33 @@ describe('Send message', () => {
             .getText()
         ).toBe('A third message');
       });
+    });
+  });
+
+  describe('Display message without contact', () => {
+    it('can display messages without contact', () => {
+      common.goToMessages();
+
+      openSendMessageModal();
+      enterCheckAndSelect(ALICE.name, 2, contactNameSelector, ALICE.name);
+      element(by.css('#send-message textarea')).sendKeys(smsMsg('contact'));
+      sendMessage();
+      clickLhsEntry(ALICE._id, ALICE.name);
+
+      browser.wait(() => {
+        return utils.deleteDocs(CONTACTS.map(contact => contact._id));
+      });
+
+      browser.wait(() => {
+        const el = element(by.css('#message-header .name'));
+        if (helper.waitUntilReady(el)) {
+          return helper.getTextFromElement(el).then(text => {
+            return text === 'Unknown sender';
+          });
+        }
+      }, 12000);
+
+      expect(getElementText('#message-header .phone')).toBe(ALICE.phone);
     });
   });
 });
